MS44 Custom hook for NavBar
