+++ app/src/Components/Layout/Navbar.tsx
@@ -2,38 +2,17 @@ import * as React from "react";
 import AppBar from "@mui/material/AppBar";
 import Box from "@mui/material/Box";
 import Toolbar from "@mui/material/Toolbar";
-import Typography from "@mui/material/Typography";
 import AccountCircleIcon from "@mui/icons-material/AccountCircle";
-import { useNavigate } from "react-router-dom";
 import { makeStyles } from "@mui/styles";
 import { useTranslation } from "react-i18next";
 import { NavBarButton } from "./NavBarButton";
-import { routes } from "../../Constants/Routes/Routes";
+import useNavBar from "../../Hooks/useNavBar";
 
 const NavBar: React.FC = () => {
-  const navigate = useNavigate();
   const styles = useStyles();
   const { t } = useTranslation("navbar");
-
-  const pages = [t("Muzik"), 
-    // t("Merch")
-];
-
-  const handleNavbarNavigation = (page: string) => {
-    if (page === t("Muzik")) {
-      navigate(routes.music);
-    } else {
-      console.log(`Page URL not found for: ${page}`);
-    }
-  };
-
-  const navigateMyProfile = React.useCallback(() => {
-    navigate(routes.myProfile);
-  }, [navigate]);
-
-  const navigateHome = React.useCallback(() => {
-    navigate(routes.home);
-  }, [navigate]);
+  const pages = [t("Muzik")];
+  const {handleNavbarNavigation, navigateHome, navigateMyProfile} = useNavBar()
 
   return (
     <AppBar className={styles.navBar} sx={{ backgroundColor: "#858585" }}>
